1. fix LC issuance pagg; 2. create specs and plan for SWIFT validation and generation

+++ runtime/component/TradeFinance/src/test/groovy/EndToEndImportLcSpec.groovy
@@ -1,3 +1,4 @@
+package trade
 
 import spock.lang.Specification
 import org.moqui.Moqui
@@ -9,19 +10,24 @@ import org.moqui.context.ExecutionContext
 class EndToEndImportLcSpec extends Specification {
     protected ExecutionContext ec
     
+    // Test constants to avoid hardcoding
+    static final String TEST_MAKER = "trade.maker"
+    static final String TEST_FACILITY_ID = "E2E-FAC-1"
+    static final String TEST_TRANS_REF = "TF-E2E-001"
+    
     def setup() {
         ec = Moqui.getExecutionContext()
-        ec.user.internalLoginUser("trade.maker")
+        ec.user.internalLoginUser(TEST_MAKER)
         ec.artifactExecution.disableAuthz()
         
-        if (ec.entity.find("moqui.security.UserAccount").condition("username", "trade.maker").count() == 0) {
+        if (ec.entity.find("moqui.security.UserAccount").condition("username", TEST_MAKER).count() == 0) {
             ec.entity.makeValue("moqui.security.UserAccount")
-                .setAll([userId: "trade.maker", username: "trade.maker", currentPassword: "trade123", firstName: "Trade", lastName: "Maker"])
+                .setAll([userId: TEST_MAKER, username: TEST_MAKER, currentPassword: "trade123", firstName: "Trade", lastName: "Maker"])
                 .create()
         }
-        if (ec.entity.find("trade.UserAuthorityProfile").condition("userId", "trade.maker").count() == 0) {
+        if (ec.entity.find("trade.UserAuthorityProfile").condition("userId", TEST_MAKER).count() == 0) {
             ec.entity.makeValue("trade.UserAuthorityProfile")
-                .setAll([authorityProfileId: "T1-E2E", userId: "trade.maker", authorityTierEnumId: "TIER_1", maxApprovalAmount: 1000000.00, currencyUomId: "USD"])
+                .setAll([userAuthorityId: "T1-E2E", userId: TEST_MAKER, delegationTierId: "TIER_1", customLimit: 1000000.00, currencyUomId: "USD", makerCheckerFlag: "MAKER_CHECKER"])
                 .create()
         }
     }
@@ -33,14 +39,14 @@ class EndToEndImportLcSpec extends Specification {
     def "Full Flow: Create LC -> Update Limit -> Generate SWIFT"() {
         setup: "Initialize Facility"
         ec.entity.makeValue("trade.CustomerFacility")
-            .setAll([facilityId:"E2E-FAC-1", totalApprovedLimit: 100000.0, utilizedAmount: 0.0]).create()
+            .setAll([facilityId: TEST_FACILITY_ID, totalApprovedLimit: 100000.0, utilizedAmount: 0.0]).create()
             
         when: "1. Create Import LC"
         def createResult = ec.service.sync().name("trade.importlc.ImportLcServices.create#ImportLetterOfCredit").parameters([
-            transactionRef: "TF-E2E-001",
+            transactionRef: TEST_TRANS_REF,
             lcAmount: 50000.0,
             lcCurrencyUomId: "USD",
-            customerFacilityId: "E2E-FAC-1",
+            customerFacilityId: TEST_FACILITY_ID,
             businessStateId: "LC_DRAFT"
         ]).call()
         def instrumentId = createResult.instrumentId
@@ -50,10 +56,10 @@ class EndToEndImportLcSpec extends Specification {
         ec.entity.find("trade.importlc.ImportLetterOfCredit").condition("instrumentId", instrumentId).one() != null
         
         when: "2. Update Limit Utilization"
-        ec.service.sync().name("trade.LimitServices.update#Utilization").parameters([facilityId:"E2E-FAC-1", amountDelta: 50000.0]).call()
+        ec.service.sync().name("trade.LimitServices.update#Utilization").parameters([facilityId: TEST_FACILITY_ID, amountDelta: 50000.0]).call()
         
         then: "Facility utilization is updated"
-        ec.entity.find("trade.CustomerFacility").condition("facilityId", "E2E-FAC-1").one().utilizedAmount == 50000.0
+        ec.entity.find("trade.CustomerFacility").condition("facilityId", TEST_FACILITY_ID).one().utilizedAmount == 50000.0
         
         when: "3. Generate SWIFT MT700"
         def swiftResult = ec.service.sync().name("trade.SwiftGenerationServices.generate#Mt700").parameters([instrumentId: instrumentId]).call()
@@ -61,13 +67,15 @@ class EndToEndImportLcSpec extends Specification {
         then: "SWIFT message is created with correct content"
         swiftResult.swiftMessageId != null
         def message = ec.entity.find("trade.importlc.SwiftMessage").condition("swiftMessageId", swiftResult.swiftMessageId).one()
-        message.messageContent.contains("TF-E2E-001")
+        message.messageContent.contains(TEST_TRANS_REF)
         message.messageType == "MT700"
         
         cleanup:
-        ec.entity.find("trade.importlc.SwiftMessage").condition("instrumentId", instrumentId).deleteAll()
-        ec.entity.find("trade.importlc.ImportLetterOfCredit").condition("instrumentId", instrumentId).deleteAll()
-        ec.entity.find("trade.TradeInstrument").condition("instrumentId", instrumentId).deleteAll()
-        ec.entity.find("trade.CustomerFacility").condition("facilityId", "E2E-FAC-1").deleteAll()
+        if (instrumentId) {
+            ec.entity.find("trade.importlc.SwiftMessage").condition("instrumentId", instrumentId).deleteAll()
+            ec.entity.find("trade.importlc.ImportLetterOfCredit").condition("instrumentId", instrumentId).deleteAll()
+            ec.entity.find("trade.TradeInstrument").condition("instrumentId", instrumentId).deleteAll()
+        }
+        ec.entity.find("trade.CustomerFacility").condition("facilityId", TEST_FACILITY_ID).deleteAll()
     }
 }
